implementation/matcha_tts/requirements.txt
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0
numpy>=1.24.0
librosa>=0.10.1
soundfile>=0.12.1
tqdm>=4.66.0
pyyaml>=6.0.1
phonemizer>=3.2.1
